EncoderH264: fix conversion warning

+++ vk_video_encoder/libs/VkVideoEncoder/VkVideoEncoderH264.cpp
@@ -331,7 +331,7 @@ VkResult VkVideoEncoderH264::ProcessDpb(VkSharedBaseObj<VkVideoEncodeFrameInfo>&
                                                    &m_h264.m_spsInfo, &pFrameInfo->stdSliceHeader,
                                                    &pFrameInfo->stdReferenceListsInfo, MAX_MEM_MGMNT_CTRL_OPS_COMMANDS);
     if (targetDpbSlot >= VkEncDpbH264::MAX_DPB_SLOTS) {
-        targetDpbSlot = (encodeFrameInfo->setupImageResource!=nullptr) + refLists.refPicListCount[0] + refLists.refPicListCount[1] + 1;
+        targetDpbSlot = static_cast<int8_t>((encodeFrameInfo->setupImageResource!=nullptr) + refLists.refPicListCount[0] + refLists.refPicListCount[1] + 1);
     }
     if (isReference) {
         assert(targetDpbSlot >= 0);
